# Fix multiple sources (performance)