--- a/.paket/Paket.Restore.targets
+++ b/.paket/Paket.Restore.targets
@@ -38,7 +38,7 @@
         <Version>%(PaketReferencesFileLinesInfo.PackageVersion)</Version>
       </PackageReference>
     </ItemGroup>
-    
+
     <PropertyGroup>
       <PaketCliToolFilePath>$(MSBuildProjectDirectory)/obj/$(MSBuildProjectFile).paket.clitools</PaketCliToolFilePath>
     </PropertyGroup>