--- a/build.sh
+++ b/build.sh
@@ -29,3 +29,4 @@ else
   fi
   mono packages/build/FAKE/tools/FAKE.exe $@ --fsiargs -d:MONO build.fsx 
 fi
+