--- a/src/Paket.Core/Common/Async.fs
+++ b/src/Paket.Core/Common/Async.fs
@@ -14,7 +14,20 @@ module AsyncExtensions =
   open System
   open System.Threading.Tasks
   open System.Threading
-
+  open System.Runtime.ExceptionServices
+  
+  // This uses a trick to get the underlying OperationCanceledException
+  let inline getCancelledException (completedTask:Task) (waitWithAwaiter) =
+      let fallback = new TaskCanceledException(completedTask) :> OperationCanceledException
+      // sadly there is no other public api to retrieve it, but to call .GetAwaiter().GetResult().
+      try waitWithAwaiter()
+          // should not happen, but just in case...
+          fallback
+      with
+      | :? OperationCanceledException as o -> o
+      | other ->
+          // shouldn't happen, but just in case...
+          new TaskCanceledException(fallback.Message, other) :> OperationCanceledException
   type Microsoft.FSharp.Control.Async with 
      /// Runs both computations in parallel and returns the result as a tuple.
     static member Parallel (a : Async<'a>, b : Async<'b>) : Async<'a * 'b> =
@@ -25,6 +38,37 @@ module AsyncExtensions =
             let! b'' = b'
             return (a'',b'')
         }
+    static member AwaitTaskWithoutAggregate (task:Task<'T>) : Async<'T> =
+        Async.FromContinuations(fun (cont, econt, ccont) ->
+            let continuation (completedTask : Task<_>) =
+                if completedTask.IsCanceled then
+                    let cancelledException =
+                        getCancelledException completedTask (fun () -> completedTask.GetAwaiter().GetResult() |> ignore)
+                    econt (cancelledException)
+                elif completedTask.IsFaulted then
+                    if completedTask.Exception.InnerExceptions.Count = 1 then
+                        econt completedTask.Exception.InnerExceptions.[0]
+                    else
+                        econt completedTask.Exception
+                else
+                    cont completedTask.Result
+            task.ContinueWith(Action<Task<'T>>(continuation)) |> ignore)
+    static member AwaitTaskWithoutAggregate (task:Task) : Async<unit> =
+        Async.FromContinuations(fun (cont, econt, ccont) ->
+            let continuation (completedTask : Task) =
+                if completedTask.IsCanceled then
+                    let cancelledException =
+                        getCancelledException completedTask (fun () -> completedTask.GetAwaiter().GetResult() |> ignore)
+                    econt (cancelledException)
+                elif completedTask.IsFaulted then
+                    if completedTask.Exception.InnerExceptions.Count = 1 then
+                        econt completedTask.Exception.InnerExceptions.[0]
+                    else
+                        econt completedTask.Exception
+                else
+                    cont ()
+            task.ContinueWith(Action<Task>(continuation)) |> ignore)
+
     static member awaitTaskWithToken (fallBack:unit -> 'T) (item: Task<'T>) : Async<'T> =
         async {
             let! ct = Async.CancellationToken