--- a/src/Paket.Core/Common/Utils.fs
+++ b/src/Paket.Core/Common/Utils.fs
@@ -14,9 +14,10 @@ open Chessie.ErrorHandling
 open Paket.Domain
 open FSharp.Polyfill
 
-#if NETSTANDARD1_6
 open System.Net.Http
-#else
+open System.Threading
+
+#if !NETSTANDARD1_6
 // TODO: Activate this in .NETCore 2.0
 ServicePointManager.SecurityProtocol <- unbox 192 ||| unbox 768 ||| unbox 3072 ||| unbox 48
                                         ///SecurityProtocolType.Tls ||| SecurityProtocolType.Tls11 ||| SecurityProtocolType.Tls12 ||| SecurityProtocolType.Ssl3
@@ -384,44 +385,42 @@ let getDefaultProxyFor =
             | Some p -> if p.GetProxy uri <> uri then p else getDefault()
             | None -> getDefault())
 
-#if USE_HTTP_CLIENT
-type WebClient = HttpClient
 type HttpClient with
-    member x.DownloadFileTaskAsync (uri : Uri, filePath : string) =
+    member x.DownloadFileTaskAsync (uri : Uri, tok : CancellationToken, filePath : string) =
       async {
-        let! response = x.GetAsync(uri) |> Async.AwaitTask
+        let! response = x.GetAsync(uri, tok) |> Async.AwaitTask
         use fileStream = new FileStream(filePath, FileMode.Create, FileAccess.Write, FileShare.None)
         do! response.Content.CopyToAsync(fileStream) |> Async.AwaitTask
         fileStream.Flush()
       } |> Async.StartAsTask
-    member x.DownloadFileTaskAsync (uri : string, filePath : string) = x.DownloadFileTaskAsync(Uri uri, filePath)
+    member x.DownloadFileTaskAsync (uri : string, tok : CancellationToken, filePath : string) = x.DownloadFileTaskAsync(Uri uri, tok, filePath)
     member x.DownloadFile (uri : string, filePath : string) =
-        x.DownloadFileTaskAsync(uri, filePath).GetAwaiter().GetResult()
+        x.DownloadFileTaskAsync(uri, CancellationToken.None, filePath).GetAwaiter().GetResult()
     member x.DownloadFile (uri : Uri, filePath : string) =
-        x.DownloadFileTaskAsync(uri, filePath).GetAwaiter().GetResult()
-    member x.DownloadStringTaskAsync (uri : Uri) =
-      async {
-        let! response = x.GetAsync(uri) |> Async.AwaitTask
+        x.DownloadFileTaskAsync(uri, CancellationToken.None, filePath).GetAwaiter().GetResult()
+    member x.DownloadStringTaskAsync (uri : Uri, tok : CancellationToken) =
+      async { 
+        let! response = x.GetAsync(uri, tok) |> Async.AwaitTask
         let! result = response.Content.ReadAsStringAsync() |> Async.AwaitTask
         return result
       } |> Async.StartAsTask
-    member x.DownloadStringTaskAsync (uri : string) = x.DownloadStringTaskAsync(Uri uri)
+    member x.DownloadStringTaskAsync (uri : string, tok : CancellationToken) = x.DownloadStringTaskAsync(Uri uri, tok)
     member x.DownloadString (uri : string) =
-        x.DownloadStringTaskAsync(uri).GetAwaiter().GetResult()
+        x.DownloadStringTaskAsync(uri, CancellationToken.None).GetAwaiter().GetResult()
     member x.DownloadString (uri : Uri) =
-        x.DownloadStringTaskAsync(uri).GetAwaiter().GetResult()
+        x.DownloadStringTaskAsync(uri, CancellationToken.None).GetAwaiter().GetResult()
 
-    member x.DownloadDataTaskAsync(uri : Uri) =
+    member x.DownloadDataTaskAsync(uri : Uri, tok : CancellationToken) =
       async {
-        let! response = x.GetAsync(uri) |> Async.AwaitTask
+        let! response = x.GetAsync(uri, tok) |> Async.AwaitTask
         let! result = response.Content.ReadAsByteArrayAsync() |> Async.AwaitTask
         return result
       } |> Async.StartAsTask
-    member x.DownloadDataTaskAsync (uri : string) = x.DownloadDataTaskAsync(Uri uri)
+    member x.DownloadDataTaskAsync (uri : string, tok : CancellationToken) = x.DownloadDataTaskAsync(Uri uri, tok)
     member x.DownloadData(uri : string) =
-        x.DownloadDataTaskAsync(uri).GetAwaiter().GetResult()
+        x.DownloadDataTaskAsync(uri, CancellationToken.None).GetAwaiter().GetResult()
     member x.DownloadData(uri : Uri) =
-        x.DownloadDataTaskAsync(uri).GetAwaiter().GetResult()
+        x.DownloadDataTaskAsync(uri, CancellationToken.None).GetAwaiter().GetResult()
 
     member x.UploadFileAsMultipart (url : Uri) filename =
         let fileTemplate = 
@@ -451,39 +450,7 @@ let internal addAcceptHeader (client:HttpClient) (contentType:string) =
 let internal addHeader (client:HttpClient) (headerKey:string) (headerVal:string) =
     client.DefaultRequestHeaders.Add(headerKey, headerVal)
 
-#else
-
-type System.Net.WebClient with
-    member x.UploadFileAsMultipart (url : Uri) filename = 
-        let fileTemplate = 
-            "--{0}\r\nContent-Disposition: form-data; name=\"{1}\"; filename=\"{2}\"\r\nContent-Type: {3}\r\n\r\n"
-        let boundary = "---------------------------" + DateTime.Now.Ticks.ToString("x", System.Globalization.CultureInfo.InvariantCulture)
-        let fileInfo = (new FileInfo(Path.GetFullPath(filename)))
-        let fileHeaderBytes = 
-            System.String.Format
-                (System.Globalization.CultureInfo.InvariantCulture, fileTemplate, boundary, "package", "package", "application/octet-stream") 
-            |> Encoding.UTF8.GetBytes
-        // we use a windows-style newline rather than Environment.NewLine for compatibility
-        let newlineBytes = "\r\n" |> Encoding.UTF8.GetBytes
-        let trailerbytes = String.Format(System.Globalization.CultureInfo.InvariantCulture, "--{0}--", boundary) |> Encoding.UTF8.GetBytes
-        x.Headers.Add(HttpRequestHeader.ContentType, "multipart/form-data; boundary=" + boundary)
-        use stream = x.OpenWrite(url, "PUT")
-        stream.Write(fileHeaderBytes, 0, fileHeaderBytes.Length)
-        use fileStream = File.OpenRead fileInfo.FullName
-        fileStream.CopyTo(stream, (4 * 1024))
-        stream.Write(newlineBytes, 0, newlineBytes.Length)
-        stream.Write(trailerbytes, 0, trailerbytes.Length)
-        stream.Write(newlineBytes, 0, newlineBytes.Length) 
-        ()
-
-let internal addAcceptHeader (client:WebClient) contentType =
-    client.Headers.Add (HttpRequestHeader.Accept, contentType)
-let internal addHeader (client:WebClient) (headerKey:string) (headerVal:string) =
-    client.Headers.Add (headerKey, headerVal)
-#endif
-
 let createWebClient (url,auth:Auth option) =
-#if USE_HTTP_CLIENT
     let handler =
         new HttpClientHandler(
             UseProxy = true,
@@ -508,42 +475,10 @@ let createWebClient (url,auth:Auth option) =
     client.DefaultRequestHeaders.Add("user-agent", "Paket")
     handler.UseProxy <- true
     client
-#else
-    let client = new WebClient()
-    client.Headers.Add("User-Agent", "Paket")
-    client.Proxy <- getDefaultProxyFor url
-
-    let githubToken = Environment.GetEnvironmentVariable "PAKET_GITHUB_API_TOKEN"
-
-    match auth with
-    | Some (Credentials(username, password)) ->
-        // htttp://stackoverflow.com/questions/16044313/webclient-httpwebrequest-with-basic-authentication-returns-404-not-found-for-v/26016919#26016919
-        //this works ONLY if the server returns 401 first
-        //client DOES NOT send credentials on first request
-        //ONLY after a 401
-        //client.Credentials <- new NetworkCredential(auth.Username,auth.Password)
-
-        //so use THIS instead to send credentials RIGHT AWAY
-        let credentials = Convert.ToBase64String(Encoding.ASCII.GetBytes(username + ":" + password))
-        client.Headers.[HttpRequestHeader.Authorization] <- sprintf "Basic %s" credentials
-        client.Credentials <- new NetworkCredential(username,password)
-    
-    | Some (Token token) ->
-        client.Headers.[HttpRequestHeader.Authorization] <- sprintf "token %s" token
-
-    | None when not (isNull githubToken) ->
-        client.Headers.[HttpRequestHeader.Authorization] <- sprintf "token %s" githubToken
-
-    | None ->
-        client.UseDefaultCredentials <- true
-    client
-#endif
-
 
 #nowarn "40"
 
 open System.Diagnostics
-open System.Threading
 open System.Collections.Generic
 open System.Runtime.ExceptionServices
 
@@ -552,10 +487,11 @@ let downloadFromUrl (auth:Auth option, url : string) (filePath: string) =
     async {
         try
             use client = createWebClient (url,auth)
+            let! tok = Async.CancellationToken
             if verbose then
                 verbosefn "Starting download from '%O'" url
             use _ = Profile.startCategory Profile.Category.NuGetDownload
-            let task = client.DownloadFileTaskAsync (Uri url, filePath) |> Async.AwaitTask
+            let task = client.DownloadFileTaskAsync (Uri url, tok, filePath) |> Async.AwaitTask
             do! task
         with
         | exn ->
@@ -567,13 +503,14 @@ let getFromUrl (auth:Auth option, url : string, contentType : string) =
     async { 
         try
             use client = createWebClient(url,auth)
+            let! tok = Async.CancellationToken
             if notNullOrEmpty contentType then
                 addAcceptHeader client contentType
 
             if verbose then
                 verbosefn "Starting request to '%O'" url
             use _ = Profile.startCategory Profile.Category.NuGetRequest
-            return! client.DownloadStringTaskAsync (Uri url) |> Async.awaitTaskWithToken (fun () -> failwithf "Uri '%O' failed to respond and cancellation was requested." url)
+            return! client.DownloadStringTaskAsync (Uri url, tok) |> Async.AwaitTask
         with
         | exn -> 
             return raise <| Exception(sprintf "Could not retrieve data from '%s'" url, exn)
@@ -584,6 +521,7 @@ let getXmlFromUrl (auth:Auth option, url : string) =
     async {
         try
             use client = createWebClient (url,auth)
+            let! tok = Async.CancellationToken
             // mimic the headers sent from nuget client to odata/ endpoints
             addAcceptHeader client "application/atom+xml, application/xml"
             addHeader client "AcceptCharset" "UTF-8"
@@ -592,7 +530,7 @@ let getXmlFromUrl (auth:Auth option, url : string) =
             if verbose then
                 verbosefn "Starting request to '%O'" url
             use _ = Profile.startCategory Profile.Category.NuGetRequest
-            return! client.DownloadStringTaskAsync (Uri url) |> Async.awaitTaskWithToken (fun () -> failwithf "Uri '%O' failed to respond and cancellation was requested." url)
+            return! client.DownloadStringTaskAsync (Uri url, tok) |> Async.AwaitTask
         with
         | exn ->
             return raise <| Exception(sprintf "Could not retrieve data from '%s'" url, exn)
@@ -623,17 +561,15 @@ let safeGetFromUrl (auth:Auth option, url : string, contentType : string) =
         try
             let uri = Uri url
             use client = createWebClient (url,auth)
+            let! tok = Async.CancellationToken
 
             if notNullOrEmpty contentType then
                 addAcceptHeader client contentType
-#if NETSTANDARD1_6
-#else
-            client.Encoding <- Encoding.UTF8
-#endif
+
             if verbose then
                 verbosefn "Starting request to '%O'" uri
             use _ = Profile.startCategory Profile.Category.NuGetRequest
-            let! raw = client.DownloadStringTaskAsync(uri) |> Async.awaitTaskWithToken (fun () -> failwithf "Uri '%O' failed to respond and cancellation was requested." uri)
+            let! raw = client.DownloadStringTaskAsync(uri, tok) |> Async.awaitTaskWithToken (fun () -> failwithf "Uri '%O' failed to respond and cancellation was requested." uri)
             return SuccessResponse raw
         with
         | :? WebException as w ->
@@ -868,13 +804,13 @@ let parseKeyValuePairs (s:string) : Dictionary<string,string> =
     | exn -> 
         raise <| Exception(sprintf "Could not parse '%s' as key/value pairs." s, exn)
 
-let downloadStringSync (url : string) (client : WebClient) = 
+let downloadStringSync (url : string) (client : HttpClient) = 
     try 
         client.DownloadString url |> ok
     with _ ->
         DownloadError url |> fail 
 
-let downloadFileSync (url : string) (fileName : string) (client : WebClient) = 
+let downloadFileSync (url : string) (fileName : string) (client : HttpClient) = 
     tracefn "Downloading file from %s to %s" url fileName
     try 
         client.DownloadFile(url, fileName) |> ok