--- a/src/Paket.Core/paket.references
+++ b/src/Paket.Core/paket.references
@@ -3,8 +3,7 @@ FSharp.Core
 Mono.Cecil
 Chessie
 FSharp.Compiler.Tools
-
-
+System.Net.Http
 
 File:Globbing.fs Common\
 File:AssemblyReader.fs Common\