… src/Paket.Core/Dependencies/NuGet.fs
… src/Paket.Core/Dependencies/NuGetCache.fs
… src/Paket.Core/Dependencies/NuGetV2.fs
… src/Paket.Core/Dependencies/NuGetV3.fs
… src/Paket.Core/Dependencies/PackageResolver.fs
… src/Paket.Core/Dependencies/RemoteDownload.fs
… src/Paket.Core/Installation/InstallProcess.fs
… src/Paket.Core/Installation/UpdateProcess.fs
… src/Paket.Core/Paket.Core.fsproj
… src/Paket.Core/Versioning/PackageSources.fs
… src/Paket.preview3/app.config
… src/Paket/Paket.fsproj
… tests/Paket.Tests/Paket.Tests.fsproj
… tests/Paket.Tests/ProjectFile/UpdateFromNugetSpecs.fs