Fixes https://github.com/fsprojects/Paket/issues/2369

Main problem was using the v2 api for myget which seems to be painfully slow. We always used v2 for all sources besides nuget, even when v3 uri was specified.
@forki  was this intended?

Solves
```
source https://dotnet.myget.org/F/dotnet-core/api/v3/index.json
source https://dotnet.myget.org/F/cli-deps/api/v3/index.json
source https://www.nuget.org/api/v2/

nuget FSharp.Core >= 4.1.0
nuget FSharp.NET.Sdk >= 1.0.0
```

in

```
Performance:
 - Resolver: 46 seconds (1 runs)
    - Runtime: 8 seconds
    - Blocked (retrieving package versions): 25 seconds (6 times)
    - Blocked (retrieving package details): 12 seconds (33 times)
    - Not Blocked (retrieving package details): 36 times
    - Not Blocked (retrieving package versions): 63 times
 - Disk IO: 10 seconds
 - Average Request Time: 1 second
 - Number of Requests: 232
```

I also added other performance improvements which brought the runtime down to 4 Minutes (instead of hours), but switching to myget was the final breakthrough:

- Properly cancel requests when we no longer need the result by delegating the cancellation token
- Do `GetVersion` requests of a SINGLE source sequentially (because we should already put the fastest one first in the common case -> fewer requests)
- Switch to System.Net.Http instead of WebRequest to have a single code-base for netcore and full.
- Add messages to the install-process to make it more obvious what paket is doing
  The install-process is taking more and more time and users are wondering what paket is doing so I added some output