@mnmr 2017-08-07
I'm wondering if the issue still remains considering that Paket (5.84.0) takes 1m 40s to check for outdated packages (and it reports almost that amount of time as being blocked, whatever that means exactly).