@matthid 2017-08-08
@mnmr do you say we regressed on this depsfile? From the timing I'd say the reason are the fallbacks recently introduced but that means it should be resolved again with the fix warnings pr...

If you mean a different depsfile please open an issue to discuss...