@mnmr 2017-08-08
I'm afraid I'm not in a position where I can answer that - I only just downloaded Paket yesterday, so do not have any reference point for earlier behaviors. I just observed that it was NuGet-like slow ;) and came across this issue while trying to find out if this was expected or not (which I'm still not entirely sure about).